Moved generic code to BaseWebRTCService.py should be moved to opentera
